+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Switched to `color-eyre` for error handling in the CLI.
 - Improved full download progress bar.
 - `Some(ProgressBar::Full)` is now the default for the library.
+- Upgraded `reqwest` dependency to `0.11`.
 
 ## [v0.4.5](https://github.com/epwalsh/rust-cached-path/releases/tag/v0.4.5) - 2020-09-15